Put FeatureList to separate file for cleaner code
